CMake: support Android, including the pjsua2 JNI bindings (#5270)

+++ pjmedia/src/pjmedia-videodev/android_dev.c
@@ -1319,11 +1319,11 @@ static void JNICALL OnGetFrame2(JNIEnv *env, jobject obj,
         PJ_LOG(1,(THIS_FILE, "Unrecognized image format from Android camera2, "
                              "please report the following plane format:"));
         PJ_LOG(1,(THIS_FILE, " Planes (buf/len/row_stride/pix_stride):"
-                             " p0=%p/%ld/%d/%d p1=%p/%ld/%d/%d "
-                             "p2=%p/%ld/%d/%d",
-                             p0, p0_len, rowStride0, pixStride0,
-                             p1, p1_len, rowStride1, pixStride1,
-                             p2, p2_len, rowStride2, pixStride2));
+                             " p0=%p/%lld/%d/%d p1=%p/%lld/%d/%d "
+                             "p2=%p/%lld/%d/%d",
+                             p0, (long long)p0_len, rowStride0, pixStride0,
+                             p1, (long long)p1_len, rowStride1, pixStride1,
+                             p2, (long long)p2_len, rowStride2, pixStride2));
 
 #if 1
         /* Generic converter to I420, based on row stride & pixel stride */
